Movie Reviews - Show movie review excerpts; Refactor Details page.
A Drawer component is either visible (open) or hidden. We control this with a boolean state variable - drawerOpen.

In the browser:

Go to the home page.
Click any movie's 'More Info' button.
Click the 'Reviews' button on the movie details page. [Some movies have no reviews, so choose another movie.]
The review extracts should overlay the top part of the page.